+++ modules/AWSPowerShell/Cmdlets/CloudFormation/Basic/Test-CFNType-Cmdlet.cs
@@ -46,7 +46,7 @@ namespace Amazon.PowerShell.Cmdlets.CFN
     /// was registered. For more information, see <a href="AWSCloudFormation/latest/APIReference/API_RegisterType.html">RegisterType</a>.
     /// </para><para>
     /// Once you've initiated testing on an extension using <code>TestType</code>, you can
-    /// use <a href="https://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_DescribeType.html">DescribeType</a>
+    /// pass the returned <code>TypeVersionArn</code> into <a href="https://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_DescribeType.html">DescribeType</a>
     /// to monitor the current test status and test status description for the extension.
     /// </para><para>
     /// An extension must have a test status of <code>PASSED</code> before it can be published.